Move toolbox code to diagramtoolbox
The code to get a tool_def will be used by the DnD code for GTK4, as well as
new_item_factory. To avoid cycles, this code has to move to diagramtoolbox.
