+++ gaphor/SysML/toolbox.py
@@ -10,8 +10,8 @@ from gaphor.diagram.diagramtoolbox import (
     ToolSection,
     general_tools,
     namespace_config,
+    new_item_factory,
 )
-from gaphor.diagram.tools import new_item_factory
 from gaphor.SysML import diagramitems as sysml_items
 from gaphor.SysML.blocks.blockstoolbox import blocks
 from gaphor.SysML.requirements.requirementstoolbox import requirements
