+++ gaphor/UML/interactions/interactionstoolbox.py
@@ -5,8 +5,12 @@ from gaphas.segment import Segment
 
 from gaphor import UML
 from gaphor.core import gettext
-from gaphor.diagram.diagramtoolbox import ToolDef, ToolSection, namespace_config
-from gaphor.diagram.tools import new_item_factory
+from gaphor.diagram.diagramtoolbox import (
+    ToolDef,
+    ToolSection,
+    namespace_config,
+    new_item_factory,
+)
 from gaphor.UML import diagramitems
 from gaphor.UML.recipes import owner_package
