+++ gaphor/diagram/diagramtoolbox.py
@@ -5,17 +5,28 @@ We bind the Toolbox to a diagram. When a diagram page (tab) is switched,
 the actions bound to the toolbuttons should change as well.
 """
 
-from typing import Callable, Collection, NamedTuple, Optional, Sequence, Tuple
+from typing import (
+    Callable,
+    Collection,
+    NamedTuple,
+    Optional,
+    Sequence,
+    Tuple,
+    Type,
+    TypeVar,
+)
 
 from gaphas.item import SE
 
 from gaphor.core import gettext
-from gaphor.core.modeling import Comment, Diagram, Presentation
+from gaphor.core.modeling import Comment, Diagram, Element, Presentation
 from gaphor.diagram import general
-from gaphor.diagram.tools import new_item_factory
+from gaphor.diagram.grouping import Group
 from gaphor.UML.recipes import owner_package
 
 ItemFactory = Callable[[Diagram, Optional[Presentation]], Presentation]
+P = TypeVar("P", bound=Presentation, covariant=True)
+ConfigFuncType = Callable[[P], None]
 
 
 def default_namespace(new_item):
@@ -59,6 +70,42 @@ def tooliter(toolbox_actions: Sequence[Tuple[str, Sequence[ToolDef]]]):
     """Iterate toolbox items, regardless of section headers."""
     for name, section in toolbox_actions:
         yield from section
+
+
+def get_tool_def(modeling_language, tool_name):
+    return next(
+        t for t in tooliter(modeling_language.toolbox_definition) if t.id == tool_name
+    )
+
+
+def new_item_factory(
+    item_class: Type[Presentation],
+    subject_class: Optional[Type[Element]] = None,
+    config_func: Optional[ConfigFuncType] = None,
+):
+    """``config_func`` may be a function accepting the newly created item."""
+
+    def item_factory(diagram, parent=None):
+        if subject_class:
+            element_factory = diagram.model
+            subject = element_factory.create(subject_class)
+        else:
+            subject = None
+
+        item = diagram.create(item_class, subject=subject)
+
+        adapter = Group(parent, item)
+        if parent and adapter.can_contain():
+            item.parent = parent
+            adapter.group()
+
+        if config_func:
+            config_func(item)
+
+        return item
+
+    item_factory.item_class = item_class  # type: ignore[attr-defined]
+    return item_factory
 
 
 general_tools = ToolSection(
