+++ tests/test_comment_line_placement.py
@@ -3,8 +3,8 @@ import pytest
 from gaphor.application import Session
 from gaphor.core import Transaction
 from gaphor.core.modeling import Diagram
+from gaphor.diagram.diagramtoolbox import new_item_factory
 from gaphor.diagram.general import CommentLineItem
-from gaphor.diagram.tools import new_item_factory
 from gaphor.diagram.tools.placement import (
     PlacementState,
     on_drag_begin,
